refactor(messages): create RichText component, unify channel/DM rendering
Move all rendering logic from StreamingMarkdown into a new RichText
component in the messages package — the canonical home for message
rendering. StreamingMarkdown becomes a one-line re-export shim so AI
components need zero changes.

Replace the weaker MessagePartRenderer (plain text, no markdown) in DM,
thread, and quote-block surfaces with RichText so channels and DMs now
render mentions, auto-linked URLs, and markdown consistently.

Also fixes stale test expectations that used the old mention:// protocol
instead of the /mention/ relative path introduced in c854613.
