+++ apps/web/src/app/dashboard/dms/[conversationId]/page.tsx
@@ -16,7 +16,7 @@ import type { FileAttachment } from '@/hooks/useAttachmentUpload';
 import { MessageAttachment } from '@/components/shared/MessageAttachment';
 import { MessageReactions, type Reaction } from '@/components/shared/MessageReactions';
 import { MessageHoverToolbar } from '@/components/shared/MessageHoverToolbar';
-import { renderMessageParts, convertToMessageParts } from '@/components/messages/MessagePartRenderer';
+import { RichText, addHardLineBreaks } from '@/components/messages/RichText';
 import type { AttachmentMeta } from '@/lib/attachment-utils';
 import useSWR from 'swr';
 import { toast } from 'sonner';
@@ -509,7 +509,7 @@ export default function InboxDMPage() {
           </div>
           {m.content && (
             <div className="text-sm break-words [overflow-wrap:anywhere]">
-              {renderMessageParts(convertToMessageParts(m.content))}
+              {<RichText content={addHardLineBreaks(m.content)} />}
             </div>
           )}
           <MessageAttachment message={m} />
@@ -687,7 +687,7 @@ export default function InboxDMPage() {
                         )}
                         {message.content && (
                           <div className="prose prose-sm dark:prose-invert max-w-none break-words [overflow-wrap:anywhere] min-w-0">
-                            {renderMessageParts(convertToMessageParts(message.content))}
+                            {<RichText content={addHardLineBreaks(message.content)} />}
                           </div>
                         )}
                         <MessageAttachment message={message} />
